+++ PerseusPluginLib/Rearrange/ChangeColumnType.cs
@@ -330,8 +330,8 @@ namespace PerseusPluginLib.Rearrange{
 			bool[,] newIsImputed = new bool[mdata.RowCount,mdata.ColumnCount + num.Length];
 			for (int i = 0; i < mdata.RowCount; i++){
 				for (int j = 0; j < mdata.ColumnCount; j++){
-					newExp[i, j] = mdata.Values.Get(i, j);
-					newQual[i, j] = mdata.Quality.Get(i, j);
+					newExp[i, j] = (float)mdata.Values.Get(i, j);
+					newQual[i, j] = (float)mdata.Quality.Get(i, j);
 					newIsImputed[i, j] = mdata.IsImputed[i, j];
 				}
 				for (int j = 0; j < newEx.Length; j++){
@@ -374,8 +374,8 @@ namespace PerseusPluginLib.Rearrange{
 			bool[,] newIsImputed = new bool[mdata.RowCount,mdata.ColumnCount + str.Length];
 			for (int i = 0; i < mdata.RowCount; i++){
 				for (int j = 0; j < mdata.ColumnCount; j++){
-					newExp[i, j] = mdata.Values.Get(i, j);
-					newQual[i, j] = mdata.Quality.Get(i, j);
+					newExp[i, j] = (float)mdata.Values.Get(i, j);
+					newQual[i, j] = (float)mdata.Quality.Get(i, j);
 					newIsImputed[i, j] = mdata.IsImputed[i, j];
 				}
 				for (int j = 0; j < newEx.Length; j++){
